# openapi.yaml
openapi: 3.1.0
info:
  title: AI PR Risk Gate API
  version: 1.0.0
  description: |
    Risk scoring and policy-gating API for pull requests.
servers:
  - url: https://ai-pr-risk-gate.onrender.com
  - url: http://localhost:8787
tags:
  - name: Health
  - name: Analysis
  - name: Analytics
  - name: Webhooks
paths:
  /health:
    get:
      tags: [Health]
      summary: Service health check
      responses:
        '200':
          description: Healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  service: { type: string }
  /health/live:
    get:
      tags: [Health]
      summary: Liveness probe
      responses:
        '200':
          description: Process is live
  /health/ready:
    get:
      tags: [Health]
      summary: Readiness probe (includes database readiness)
      responses:
        '200':
          description: Ready
        '503':
          description: Not ready
  /api/analyze:
    post:
      tags: [Analysis]
      summary: Analyze PR changes and apply policy gate
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnalyzeRequest'
      responses:
        '200':
          description: Analysis complete, policy allows
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AnalyzeResponse' }
        '409':
          description: Analysis complete, policy blocks
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AnalyzeResponse' }
        '400':
          description: Invalid payload
  /api/trends:
    get:
      tags: [Analytics]
      summary: Daily trend of average risk score
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      parameters:
        - in: query
          name: repo
          schema: { type: string }
        - in: query
          name: days
          schema: { type: integer, minimum: 1, maximum: 365, default: 30 }
      responses:
        '200':
          description: Trend data
  /api/recent:
    get:
      tags: [Analytics]
      summary: Most recent assessments
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      parameters:
        - in: query
          name: repo
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      responses:
        '200':
          description: Recent rows
  /api/severity:
    get:
      tags: [Analytics]
      summary: Severity distribution over date range
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      parameters:
        - in: query
          name: repo
          schema: { type: string }
        - in: query
          name: days
          schema: { type: integer, minimum: 1, maximum: 365, default: 30 }
      responses:
        '200':
          description: Severity counts
  /api/findings:
    get:
      tags: [Analytics]
      summary: Top recurring findings over date range
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      parameters:
        - in: query
          name: repo
          schema: { type: string }
        - in: query
          name: days
          schema: { type: integer, minimum: 1, maximum: 365, default: 30 }
      responses:
        '200':
          description: Finding counts
  /api/onboarding/issue-key:
    post:
      tags: [Analysis]
      summary: Issue a scoped write API key (self-serve onboarding)
      description: Disabled by default; when enabled, issues a repo-scoped write key with expiry.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [repo]
              properties:
                repo: { type: string }
                ownerLabel: { type: string }
      responses:
        '201': { description: Key issued }
        '400': { description: Invalid request }
        '403': { description: Repo not allowed }
        '404': { description: Onboarding disabled }
        '429': { description: Rate limit exceeded }
  /webhook/github:
    post:
      tags: [Webhooks]
      summary: GitHub pull_request webhook endpoint
      description: Validates X-Hub-Signature-256, fetches PR files, analyzes risk, comments on PR. Can also be authorized with a write API key via X-API-Key or Bearer token.
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      parameters:
        - in: header
          name: X-Hub-Signature-256
          required: false
          schema: { type: string }
      responses:
        '202': { description: Accepted / analysis complete and allowed }
        '409': { description: Analysis complete and blocked by policy }
        '401': { description: Invalid signature }
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: token
  schemas:
    ChangedFile:
      type: object
      required: [filename]
      properties:
        filename: { type: string }
        status: { type: string }
        patch: { type: string }
    AnalyzeRequest:
      type: object
      required: [repo, prNumber, files]
      properties:
        repo: { type: string }
        owner: { type: string }
        prNumber: { type: integer }
        files:
          type: array
          items: { $ref: '#/components/schemas/ChangedFile' }
    AnalyzeResponse:
      type: object
      properties:
        score: { type: integer }
        severity: { type: string, enum: [low, medium, high, critical] }
        findings:
          type: array
          items: { type: string }
        recommendations:
          type: array
          items: { type: string }
        policy:
          type: object
          properties:
            allowed: { type: boolean }
            reason: { type: string }